Flask/requirements.txt
requests
flask
SQLAlchemy==1.4.49
psycopg2-binary
python_dotenv
pydantic==1.10.9